==> 02_split_by_publication_date.ipynb <==
import pandas as pd
from os.path import join

dest = pd.read_csv("data/csv/2023-04-02-destombes.csv.gz")

dest.columns

dest = pd.read_csv("data/csv/2023-04-02-destombes.csv.gz")
print(len(dest))
dest = dest[~dest['publication_date'].isna()]
dest = dest[dest['publication_date'].str.match('\d{4}')]
print(len(dest))
dest['publication_date'] = dest['publication_date'].astype(int)
dest = dest[dest['publication_date'] >= 1880]
dest = dest[dest['publication_date'] < 1890]
print(len(dest))

dest.to_csv("data/csv/2023-04-02-destombes_1880.csv.gz", index=False)

folder = "/media/fpichenot/Samsung_T5/bnf_data/SRU/csv"
bnf_mon_1875_1894 = pd.DataFrame()
nb_not = 0
for year in range(1875, 1895):
    bnf_mon = pd.read_csv(join(folder, f"bnf_mon_{year}.csv"))
    print(len(bnf_mon))
    nb_not = nb_not + len(bnf_mon)
    bnf_mon_1875_1894 = pd.concat([bnf_mon_1875_1894, bnf_mon])
print(nb_not)

len(bnf_mon_1875_1894)

bnf_mon_1875_1894.to_csv(join("data/csv", "bnf_mon_1875_1894.csv.gz"), index=False)
